refactor: tasks routes with authorization for each user

+++ src/tasks/tasks.service.ts
@@ -14,8 +14,14 @@ export class TasksService {
     private _taskRepository: TaskRepository,
   ) {}
 
-  async getTaskById(id: number): Promise<Task> {
-    const taskFound = await this._taskRepository.findOne(id);
+  async getTasks(filterDto: GetTasksFilterDto, user: User): Promise<Task[]> {
+    return this._taskRepository.getTasks(filterDto, user);
+  }
+
+  async getTaskById(id: number, user: User): Promise<Task> {
+    const taskFound = await this._taskRepository.findOne({
+      where: { id, userId: user.id },
+    });
     if (!taskFound) {
       throw new NotFoundException(`Task with ID ${id} was not found`);
     }
@@ -26,21 +32,21 @@ export class TasksService {
     return await this._taskRepository.createTask(createTaskDto, user);
   }
 
-  async updateTaskStatus(id: number, status: TaskStatus): Promise<Task> {
-    const task = await this.getTaskById(id);
+  async updateTaskStatus(
+    id: number,
+    status: TaskStatus,
+    user: User,
+  ): Promise<Task> {
+    const task = await this.getTaskById(id, user);
     task.status = status;
     await task.save();
     return task;
   }
 
-  async deleteTask(id: number): Promise<void> {
-    const result = await this._taskRepository.delete(id);
+  async deleteTask(id: number, user: User): Promise<void> {
+    const result = await this._taskRepository.delete({ id, userId: user.id });
     if (result.affected === 0) {
       throw new NotFoundException(`Task with ID ${id} was not found`);
     }
-  }
-
-  async getTasks(filterDto: GetTasksFilterDto): Promise<Task[]> {
-    return this._taskRepository.getTasks(filterDto);
   }
 }
